Fix usage, clarify comments

+++ run_adapt.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -15,22 +14,30 @@
 # limitations under the License.
 
 # This script constructs a new phoneme-based lexicon based on a supplied dictionary
-# file (defaults to cantab-TEDLIUM.dct)
+# file (defaults to cantab-TEDLIUM.dct), a list of "new words" to add to the dict,
+# and a training text
+# then composes the resulting Language, Grammar, and Token graphs into a new
+# decoding graph suitable for use decoding speech with eesen-transcriber
+
+# required inputs: newwords.txt, example_txt
 
 dict=../db/cantab-TEDLIUM/cantab-TEDLIUM.dct
-#dict=../db/cantab-TEDLIUM/TEDLIUM.152k.dic
 lm_folder=adapt_lm
 langdir=data/lang_phn
 
 # Generate pronunciations for words in supplied text file "newwords.txt"
 
 # Get pronunciations from CMU lex tool (requires internet connectivity!)
-curl -s `curl -s -F "wordfile=@newwords.txt" http://www.speech.cs.cmu.edu/cgi-bin/tools/logios/lextool.pl | awk ' /DICT/ { print $3 } '` | sed 's/ TS$/ T S/' | sed 's/ IX/ EH/g' | awk '{printf tolower($1 " "); for (i=2; i<=NF; i++)  
+# the last bit fixes case: downcase word, leave phonemes upcased
+if [ -f newwords.txt ]; then
+  curl -s `curl -s -F "wordfile=@newwords.txt" http://www.speech.cs.cmu.edu/cgi-bin/tools/logios/lextool.pl | awk ' /DICT/ { print $3 } '` | sed 's/(.)//g' | awk '{printf tolower($1 " "); for (i=2; i<=NF; i++)  
    printf ($i " "); print ""}' > newwords.dct
-cat newwords.dct $dict | sort > newdict.dct
+  cat newwords.dct $dict | sort > newdict.dct
+else
+  cp $dict newdict.dct
+fi
 
-
-# Construct the phoneme-based lexicon
+# Construct the phoneme-based lexicon, output in data/local/dict_phn
 echo "Constructing the phoneme-based lexicon"
 ../local/tedlium_prepare_phn_dict.sh --srcdict newdict.dct || exit 1;
 
@@ -42,12 +49,7 @@ echo "Compiling the lexicon and token FSTs"
 # train_lms.sh takes lang_bd/words.txt and wordlist.txt as input; create anew from dictionary
 # creates output file adapt_lm/3gram-mincount/lm_pr6.0.gz
 
-# why doesn't this work?
-#cat $langdir/words.txt | awk '{print $1}' | sort | uniq > wordlist.txt
 cat newdict.dct | awk '{print $1}' | sort | uniq > wordlist.txt
-
-# this works
-#cat $dict | awk '{print $1}' | sort | uniq > wordlist.txt
 
 cp $langdir/words.txt lang_bd
 # lingering data corrupts output, so clear first
@@ -65,4 +67,4 @@ gunzip -c $arpa_lm | utils/find_arpa_oovs.pl $langdir/words.txt > $oov_list || e
 echo "Composing the decoding graph using our own ARPA LM"
 ../local/tedlium_decode_graph.sh --arpa_lm $arpa_lm --oov_list $oov_list $langdir || exit 1;
 
-echo "Finished adapting LM"
+echo "Finished adapting LM, results in " `pwd`/${langdir}_test
